+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+### Added
+- Added script to look up for log4j2 dangerous classes and remove them because of a security issue (CVE-2021-44228).
+
 ## [2.0.1] - 2021-07-07
 ### Changed
 - Allow override of `hive.metastore.disallow.incompatible.col.type.changes=true` property.